refactoring actions WIP
