+++ frontend/src/reducers/commentReducer.js
@@ -24,10 +24,9 @@ export default function commentReducer(
             let newStateBody = []
             action.payload.map(comment => state.body.find(i => i.id == comment.id) ? null : newStateBody.push(comment))
             state.body.map(i => newStateBody.push(i))
-            debugger
+            
         return{
             ...state, loading: false, loaded: true, bulkLoad: true, body: [...newStateBody]
-            // ...state, loading: false, loaded: true, bulkLoad: true, body: [...state.body, ...action.payload]
         }
 
         case 'LOAD_COMMUNITY_DATA':
@@ -35,9 +34,9 @@ export default function commentReducer(
             let newCommunityStateBody = []
             action.payload.comments.map(comment => state.body.find(i => i.id == comment.id) ? null : newCommunityStateBody.push(comment))
             state.body.map(i => newCommunityStateBody.push(i))
+
             return{
                 ...state, loading: false, loaded: true, bulkload: true, body: [ ...newCommunityStateBody]
-                // ...state, loading: false, loaded: true, bulkload: true, body: [...state.body, ...action.payload.comments]
             }
 
         case 'LOAD_NEW_COMMENT':
@@ -45,7 +44,7 @@ export default function commentReducer(
             ...state, loading: false, loaded: true, body: [...state.body, action.payload]
         }  
 
-        case 'LOAD_NEW_COMMUNIT_COMMENT':
+        case 'LOAD_NEW_COMMUNITY_COMMENT':
             return{
                 ...state, loading: false, loaded: true, body: [...state.body, action.payload]
             }
